cmd/kraft/build: Migrate `menuconfig` to root command and rename
